requirements.txt
discord.py>=2.4.0
python-dotenv>=1.0.1
aiofiles>=23.2.1
aiosqlite>=0.20.0
supabase>=2.0.0